requirements.txt
tornado
requests
websocket-client
pyqrcode
pypng
